+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.util.MainActivity"
             android:exported="false"/>
         <activity
-            android:name=".SplashActivity"
+            android:name=".ui.util.SplashActivity"
             android:theme="@style/Splash"
             android:exported="true">
             <intent-filter>
