--- app/src/main/java/com/example/ecomate/SignupActivity.kt
@@ -1,14 +1,0 @@
-package com.example.ecomate
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.ecomate.databinding.ActivitySignupBinding
-
-class SignupActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val binding = ActivitySignupBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-    }
-}
