+++ app/src/main/java/com/example/ecomate/ui/adapter/ChallengeAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.ecomate.adapter
+package com.example.ecomate.ui.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
