+++ app/src/main/java/com/example/ecomate/ui/community/PostAddActivity.kt
@@ -1,13 +1,13 @@
-package com.example.ecomate
+package com.example.ecomate.ui.community
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.ecomate.databinding.ActivityPostInfoBinding
+import com.example.ecomate.databinding.ActivityPostAddBinding
 
-class PostInfoActivity : AppCompatActivity() {
+class PostAddActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val binding = ActivityPostInfoBinding.inflate(layoutInflater)
+        val binding = ActivityPostAddBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
     }
