+++ app/src/main/java/com/example/ecomate/ui/map/MapFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ecomate.fragment
+package com.example.ecomate.ui.map
 
 import android.os.Bundle
 import android.view.LayoutInflater
