+++ app/src/main/java/com/example/ecomate/ui/util/PermissionActivity.kt
@@ -1,4 +1,4 @@
-package com.example.ecomate
+package com.example.ecomate.ui.util
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
