+++ app/src/main/java/com/example/ecomate/ui/util/SplashActivity.kt
@@ -1,9 +1,10 @@
-package com.example.ecomate
+package com.example.ecomate.ui.util
 
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import androidx.appcompat.app.AppCompatActivity
+import com.example.ecomate.R
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
